refactor(ios): split chat content models out of MobileChatContentParser
Behavior-preserving move (MobileChatContentParser.swift 1,327 -> 1,210
lines): the 10 wire-format model types (MobileChatToolCallState,
MobileChatToolCallCardModel, MobileChatEntityKind, MobileChatProseRun,
MobileChatMerchArtifact, MobileChatMerchOptionsPayload,
MobileChatMerchOptionCard, MobileChatMerchDesignsPayload,
MobileChatMerchDesignCard, MobileChatRenderableSegment) move verbatim
to Core/MobileChatContentModels.swift. Parser, tokenizer, and label
tables stay put (next chunk). pbxproj wired with explicit entries.
